requirements.txt
appnope==0.1.0
attrs==20.3.0
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
decorator==4.4.2
geojson==2.5.0
geojson-utils==0.0.2
idna==2.10
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
munch==2.5.0
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
requests==2.25.0
Shapely==1.7.1
six==1.15.0
text-unidecode==1.3
topojson==1.0rc11
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
